get pytest and mocking pynamodb working on meta table;
